Docker/requirements.txt
gradio
python-dotenv
langchain-openai 
qdrant-client
langchain-qdrant
unstructured==0.7.12 
pinecone-client
openai 
langchain 
langchain-pinecone 
langchain-community
wikipedia
pypdf
sentence-transformers
langchain-huggingface
clean-text
torch
# torchvision
# torchaudio
xformers==0.0.27